perf: Lazily pre-populates cache for CompatibleMaterial#getFurnaceResult
The previous implementation would not cache the result for Materials without a furnace result.
So it would always check all recepies before return null.

We now just populate the whole cache instead of populating each Material<->Result combination one-by-one.
We are already iterating all the recepies anyways and EpicFurnaces should be the only plugin to use this method.
So there's actually not even a memory-consumtion trade-off right now because over time it would be
like this anyway – Just with worse performance